refactor(subagent-runner): extract single runner for in-process + facet subagents
Pre-refactor, the explore and task subagents had two parallel implementations:
agentic.ts had buildExploreTool / buildTaskTool (in-process path), and
explore-agent.ts / task-agent.ts had the facet DOs (out-of-process path).
Both paths called generateText with the same prompts, step caps, prepareStep,
and tool subset — but had already started to drift on prompt text and timeouts.

This commit:

1. Creates src/subagent-runner.ts as the single source of truth. Moves
   EXPLORE_MAX_STEPS, TASK_TIMEOUT_MS, the system prompts, capToolOutputs,
   subagentPrepareStep, resolveSubagentModel, and buildProviderForModel
   out of agentic.ts.

2. Reduces agentic.ts's public surface to buildProvider + buildToolsForThink.
   The 17 internal helpers that only existed for explore-agent.ts and
   task-agent.ts to reach are gone.

3. Both entry points (in-process tools, facet DO methods) now call
   runSubagent(...). Drift is impossible by construction.

4. Fixes the AGENTS.md invariant violation in explore-agent.ts and
   task-agent.ts (direct imports from @cloudflare/think/tools/workspace).
   think-adapter.ts now re-exports the individual workspace tool factories.

Per the improve-codebase-architecture audit, candidate #1.
