+++ test/facet-explore-parallel-unit.test.ts
@@ -27,6 +27,7 @@ vi.mock("@cloudflare/codemode", async (importOriginal) => {
   };
 });
 vi.mock("../src/agentic", async () => await import("./helpers/agentic-mock"));
+vi.mock("../src/subagent-runner", async () => await import("./helpers/subagent-runner-mock"));
 vi.mock("../src/notify", () => ({
   sendNotification: vi.fn(),
 }));
