+++ test/facet-task-scratch-unit.test.ts
@@ -30,6 +30,7 @@ vi.mock("@cloudflare/codemode", async (importOriginal) => {
   };
 });
 vi.mock("../src/agentic", async () => await import("./helpers/agentic-mock"));
+vi.mock("../src/subagent-runner", async () => await import("./helpers/subagent-runner-mock"));
 vi.mock("../src/notify", () => ({
   sendNotification: vi.fn(),
 }));
